[add - datastore_client] DatastoreClient をエンティティごとにサブクラス化するように変更
